examples/rice_diseases/requirements.txt
scikit-image>=0.19.0
networkx>=2.6
matplotlib>=3.5.0
pillow>=9.0.0
tqdm>=4.62.0
scipy>=1.7.0
